requirements_linux.txt
rich
mdurl
wcwidth
Pygments
python-magic
markdown-it-py
typing_extensions